修复农夫乐事

+++ src/main/java/com/mohistmc/mod/module/farmersdelight/common/block/RiceBaleBlock.java
@@ -26,7 +26,7 @@ public class RiceBaleBlock extends Block
 
 	@Override
 	public void fallOn(Level level, BlockState state, BlockPos pos, Entity entity, double fallDistance) {
-		entity.causeFallDamage(fallDistance, 0.2F, level.damageSources().fall());
+		entity.causeFallDamage((float) fallDistance, 0.2F, level.damageSources().fall());
 	}
 
 	@Override
